Move MediaType declarations into http module
