+++ cosmos-server/src/it/scala/com/mesosphere/cosmos/PackageDescribeSpec.scala
@@ -5,7 +5,7 @@ import cats.data.Xor.Right
 import com.mesosphere.cosmos.circe.Decoders._
 import com.mesosphere.cosmos.http.MediaTypes
 import com.mesosphere.cosmos.rpc.v1.circe.Encoders._
-import com.mesosphere.cosmos.rpc.v1.model.{DescribeRequest, ListVersionsRequest}
+import com.mesosphere.cosmos.rpc.v1.model.{DescribeRequest, ErrorResponse, ListVersionsRequest}
 import com.mesosphere.cosmos.thirdparty.marathon.model.{AppId, MarathonApp, MarathonAppContainer, MarathonAppContainerDocker}
 import com.mesosphere.cosmos.test.CosmosIntegrationTestClient
 import com.mesosphere.universe.v2.circe.Decoders._
