+++ cosmos-server/src/it/scala/com/mesosphere/cosmos/handler/PackageRenderHandlerSpec.scala
@@ -1,12 +1,11 @@
 package com.mesosphere.cosmos.handler
 
 import cats.data.Xor
-import com.mesosphere.cosmos.ErrorResponse
 import com.mesosphere.cosmos.circe.Decoders._
 import com.mesosphere.cosmos.http.MediaTypes
 import com.mesosphere.cosmos.rpc.v1.circe.Decoders._
 import com.mesosphere.cosmos.rpc.v1.circe.Encoders._
-import com.mesosphere.cosmos.rpc.v1.model.{RenderRequest, RenderResponse}
+import com.mesosphere.cosmos.rpc.v1.model.{ErrorResponse, RenderRequest, RenderResponse}
 import com.mesosphere.cosmos.test.CosmosIntegrationTestClient
 import com.mesosphere.universe.v2.model.PackageDetailsVersion
 import com.twitter.finagle.http.Status
