+++ cosmos-server/src/it/scala/com/mesosphere/cosmos/handler/UninstallHandlerSpec.scala
@@ -1,13 +1,11 @@
 package com.mesosphere.cosmos.handler
 
 import java.util.UUID
-
 import cats.data.Xor
-import com.mesosphere.cosmos.ErrorResponse
 import com.mesosphere.cosmos.circe.Decoders._
 import com.mesosphere.cosmos.http.MediaTypes
 import com.mesosphere.cosmos.rpc.v1.circe.Decoders._
-import com.mesosphere.cosmos.rpc.v1.model.UninstallResponse
+import com.mesosphere.cosmos.rpc.v1.model.{ErrorResponse, UninstallResponse}
 import com.mesosphere.cosmos.test.CosmosIntegrationTestClient
 import com.mesosphere.cosmos.thirdparty.marathon.model.AppId
 import com.netaporter.uri.dsl._
