+++ cosmos-server/src/it/scala/com/mesosphere/cosmos/rpc/v1/model/ErrorResponseSpec.scala
@@ -1,8 +1,8 @@
-package com.mesosphere.cosmos
+package com.mesosphere.cosmos.rpc.v1.model
 
 import cats.data.Xor
 import com.mesosphere.cosmos.circe.Decoders._
-import com.mesosphere.cosmos.http.MediaTypes
+import com.mesosphere.cosmos.rpc.MediaTypes
 import com.mesosphere.cosmos.test.CosmosIntegrationTestClient._
 import com.twitter.finagle.http.Status
 import com.twitter.io.Buf
