+++ cosmos-server/src/main/scala/com/mesosphere/cosmos/circe/Decoders.scala
@@ -1,8 +1,8 @@
 package com.mesosphere.cosmos.circe
 
 import com.mesosphere.universe.common.circe.Decoders._
-import com.mesosphere.cosmos.ErrorResponse
 import com.mesosphere.cosmos.model.ZooKeeperStorageEnvelope
+import com.mesosphere.cosmos.rpc.v1.model.ErrorResponse
 import io.circe.Decoder
 import io.circe.generic.semiauto._
