+++ cosmos-server/src/main/scala/com/mesosphere/cosmos/repository/UniverseClient.scala
@@ -11,12 +11,13 @@ import cats.data.Xor.{Left, Right}
 import com.mesosphere.cosmos._
 import com.mesosphere.cosmos.converter.InternalPackageDefinition._
 import com.mesosphere.cosmos.http.MediaTypeOps._
-import com.mesosphere.cosmos.http.{MediaTypeParseError, MediaTypeParser, MediaTypes, RequestSession}
+import com.mesosphere.cosmos.http.{MediaTypeParseError, MediaTypeParser, RequestSession}
 import com.mesosphere.cosmos.rpc.v1.model.PackageRepository
 import com.mesosphere.universe
 import com.mesosphere.universe.v2.circe.Decoders._
 import com.mesosphere.universe.v3.circe.Decoders._
 import com.mesosphere.universe.bijection.UniverseConversions._
+import com.mesosphere.universe.MediaTypes
 import com.netaporter.uri.Uri
 import com.twitter.bijection.Conversion.asMethod
 import com.twitter.finagle.stats.{NullStatsReceiver, Stat, StatsReceiver}
@@ -73,7 +74,7 @@ final class DefaultUniverseClient(adminRouter: AdminRouter)(implicit statsReceiv
                 Option(conn.getHeaderField("Content-Type")).getOrElse(
                   throw UnsupportedContentType(
                     List(MediaTypes.UniverseV3Repository,
-                      MediaTypes.applicationZip)
+                      MediaTypes.UniverseV2Repository)
                   )
                 )
               val parsedContentType = MediaTypeParser
@@ -82,7 +83,7 @@ final class DefaultUniverseClient(adminRouter: AdminRouter)(implicit statsReceiv
                   case MediaTypeParseError(msg, cause) =>
                     throw UnsupportedContentType(
                       List(MediaTypes.UniverseV3Repository,
-                        MediaTypes.applicationZip)
+                        MediaTypes.UniverseV2Repository)
                     )
                 }
                 .get
@@ -124,7 +125,7 @@ final class DefaultUniverseClient(adminRouter: AdminRouter)(implicit statsReceiv
                   case Xor.Right(repo) => repo
                 }
               }
-            } else if (contentType.isCompatibleWith(MediaTypes.applicationZip)) {
+            } else if (contentType.isCompatibleWith(MediaTypes.UniverseV2Repository)) {
               val v2Scope = decodeScope.scope("v2")
               v2Scope.counter("count").incr()
               Stat.time(v2Scope.stat("histogram")) {
@@ -133,7 +134,7 @@ final class DefaultUniverseClient(adminRouter: AdminRouter)(implicit statsReceiv
             } else {
               throw UnsupportedContentType.forMediaType(
                   List(MediaTypes.UniverseV3Repository,
-                       MediaTypes.applicationZip),
+                       MediaTypes.UniverseV2Repository),
                   Some(contentType)
               )
             }
