+++ cosmos-server/src/main/scala/com/mesosphere/cosmos/rpc/v1/circe/MediaTypedRequestDecoders.scala
@@ -2,7 +2,7 @@ package com.mesosphere.cosmos.rpc.v1.circe
 
 import com.mesosphere.cosmos.circe.MediaTypedDecoder
 import com.mesosphere.cosmos.finch.MediaTypedRequestDecoder
-import com.mesosphere.cosmos.http.MediaTypes
+import com.mesosphere.cosmos.rpc.MediaTypes
 import com.mesosphere.cosmos.rpc.v1.circe.Decoders._
 import com.mesosphere.cosmos.rpc.v1.model._
